move all LJ functors to md directory
